Updated agreement form around sponsorship commitment and host rotations
We want to ensure we are growing a wider tech community in Kent, as such we would like to host our core codeHarbour events across the county. As such, we want to rotate locations on a monthly basis, so we have added clarification for previous hosts as to the state of their sponsorship after the rotation.

We have also added more clarity to the commitment of sponsorship, along with rules for continuous sponsorship. Any questions give myself or one of the organisers a shout.

Cheers, Louis